make names more intuitive (days -> month) and move out Week component into its own file
